+++ src/pages/DashboardPage/DashBoardPage.module.css
@@ -2,4 +2,7 @@
   flex: 1;
   background-color: #fafafa;
   display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  justify-content: flex-start;
 }
